handles sqon with no fields specified and organizes file structures
